+++ templates/types/extractor/fastapi/app/api/models/output.py
@@ -1,6 +1,7 @@
 import logging
-from llama_index.core.schema import BaseModel, Field
 from typing import List
+
+from llama_index.core.schema import BaseModel, Field
 
 logger = logging.getLogger("uvicorn")
