+++ templates/types/extractor/fastapi/app/api/routers/__init__.py
@@ -1,0 +1,1 @@
+from .extractor import query_request
