+++ templates/types/extractor/fastapi/app/api/routers/extractor.py
@@ -1,12 +1,10 @@
 import logging
-import os
 
-from fastapi import APIRouter, HTTPException
-from llama_index.core.settings import Settings
+from fastapi import APIRouter
 from pydantic import BaseModel
 
-from app.api.routers.output import Output
-from app.engine.index import get_index
+from app.api.models.output import Output
+from app.services.engine import get_query_engine
 
 extractor_router = r = APIRouter()
 
@@ -37,22 +35,3 @@ async def query_request(
     return Output(**output_data)
 
 
-def get_query_engine(output_cls: BaseModel):
-    top_k = os.getenv("TOP_K", 3)
-
-    index = get_index()
-    if index is None:
-        raise HTTPException(
-            status_code=500,
-            detail=str(
-                "StorageContext is empty - call 'poetry run generate' to generate the storage first"
-            ),
-        )
-
-    sllm = Settings.llm.as_structured_llm(output_cls)
-
-    return index.as_query_engine(
-        similarity_top_k=int(top_k),
-        llm=sllm,
-        response_mode="tree_summarize",
-    )
