+++ templates/types/extractor/fastapi/app/ui/components/upload.py
