+++ templates/types/extractor/fastapi/app/ui/pages/__init__.py
