+++ templates/types/extractor/fastapi/app/ui/pages/index.py
