+++ templates/types/extractor/fastapi/app/ui/styles.py
