+++ templates/types/extractor/fastapi/app/ui/templates/__init__.py
