+++ templates/types/extractor/fastapi/app/ui/templates/template.py
