--- templates/types/extractor/fastapi/extractor/extractor.py
@@ -1,6 +1,0 @@
-import reflex as rx
-
-from .pages import *
-
-# Create the app.
-app = rx.App()
