+++ templates/types/extractor/fastapi/rxconfig.py
@@ -1,5 +1,5 @@
 import reflex as rx
 
 config = rx.Config(
-    app_name="extractor",
+    app_name="app",
 )
